+++ CHANGELOG.md
@@ -1,36 +1,50 @@
+# 0.1.0
+
+## Feat
+
+- Added support for custom notification builders.
+- Added support for more SnackyLocations. (top, topStart, topEnd, bottom, bottomStart, bottomEnd)
+- Better swipe to dismiss support.
+
 # 0.0.14
 
 ## Fix
+
 - Better documentation in the `README.md` about the `SnackyNavigationObserver`
 
 # 0.0.12 - 0.0.13
 
 ## Fix
+
 - Extra documentation in the `README.md`
 
 # 0.0.10 - 0.0.11
 
 ## Fix
+
 - Automated builds using GitHub Actions
 - Added preview video to the `README.md`
 
 # 0.0.4 - 0.0.9
 
 ## Fix
+
 - Added `Snacky` logo to the `README.md`
 
 # 0.0.3
 
 ## Fix
+
 - Exported all requried files
 - Cancel all snackies when a new route is pushed or a route is replaced `SnackyNavigationObserver`
 
 # 0.0.2
 
 ## Fix
+
 - Description of the package
 - Spacing between leading & text & trailing widgets
 
 # 0.0.1
 
-* initial release
+- initial release